Rename particle system source file.
